ml_framework/swin/requirements.txt
torch
timm
pandas
torchvision
pytorch_lightning
scikit-learn
torchmetrics
pillow
